Modules update: renamed apps to module, some restructuring
Following the proposals of Sunil in issue #447.
