+++ plinth/modules/deluge/__init__.py
@@ -21,102 +21,14 @@ Plinth module to configure a Deluge web client.
 
 from django.utils.translation import ugettext_lazy as _
 
-from plinth.modules import PlinthModuleManager
 from plinth import actions
 from plinth import action_utils
 from plinth import cfg
 from plinth import frontpage
 from plinth import service as service_module
 
-
-# customize the module manager name if it is different from 'ModuleManager'
-# module_manager_name = 'AnotherModuleManager'
-
-
-class ModuleManager(PlinthModuleManager):
-    """
-    Example usage of a ModuleManager (without much customization)
-    """
-    version = 1
-
-    depends = ['apps']
-
-    service = None
-
-    managed_services = ['deluge-web']
-
-    managed_packages = ['deluged', 'deluge-web']
-
-    title = _('BitTorrent Web Client \n (Deluge)')
-
-    description = [
-        _('Deluge is a BitTorrent client that features a Web UI.'),
-
-        _('When enabled, the Deluge web client will be available from '
-          '<a href="/deluge">/deluge</a> path on the web server. The '
-          'default password is \'deluge\', but you should log in and change '
-          'it immediately after enabling this service.')
-    ]
-
-    def init(self):
-        """Initialize the Deluge module."""
-        menu = cfg.main_menu.get('apps:index')
-        menu.add_urlname(title, 'glyphicon-magnet', 'deluge:index')
-
-        global service
-        setup_helper = globals()['setup_helper']
-        if setup_helper.get_state() != 'needs-setup':
-            service = service_module.Service(
-                managed_services[0], title, ports=['http', 'https'],
-                is_external=True, is_enabled=is_enabled, enable=enable,
-                disable=disable)
-
-            if is_enabled():
-                add_shortcut()
-
-    def setup(self, helper, old_version=None):
-        """Install and configure the module."""
-        helper.install(managed_packages)
-        helper.call('post', actions.superuser_run, 'deluge', ['enable'])
-        global service
-        if service is None:
-            service = service_module.Service(
-                managed_services[0], title, ports=['http', 'https'],
-                is_external=True, is_enabled=is_enabled, enable=enable,
-                disable=disable)
-        helper.call('post', service.notify_enabled, None, True)
-        helper.call('post', add_shortcut)
-
-    def add_shortcut(self):
-        frontpage.add_shortcut('deluge', title, url='/deluge',
-                               login_required=True)
-
-    def is_enabled(self):
-        """Return whether the module is enabled."""
-        return (action_utils.webserver_is_enabled('deluge-plinth') and
-                action_utils.service_is_enabled('deluge-web'))
-
-    def enable(self):
-        """Enable the module."""
-        actions.superuser_run('deluge', ['enable'])
-        # TODO: move add_shortcut functionality to default enable method
-        add_shortcut()
-
-    def disable(self):
-        """Disable the module."""
-        actions.superuser_run('deluge', ['disable'])
-        frontpage.remove_shortcut('deluge')
-
-    def diagnose(self):
-        """Run diagnostics and return the results."""
-        results = []
-
-        results.append(action_utils.diagnose_port_listening(8112, 'tcp4'))
-        results.append(action_utils.diagnose_port_listening(8112, 'tcp6'))
-        results.extend(action_utils.diagnose_url_on_all(
-            'https://{host}/deluge', check_certificate=False))
-
-        return results
+# customize the app if it is different from 'apps.App'
+# default_app = 'MyAppName'
 
 
 version = 1
